* parser.y : fix bug #108 figurative constants not accepted in screen items * testsuite.src/syn_screen.at : Added first testsuite entries for SCREEN section * minor changes for fixing compiler warnings / documentation
git-svn-id: svn://svn.code.sf.net/p/open-cobol/code/branches/gnu-cobol-2.0@457 ed166372-6744-4ac0-a67f-bb1ae9efa102

+++ libcob/intrinsic.c
@@ -1577,7 +1577,7 @@ num_leading_nonspace (const char *str)
 	int i;
 	int str_len = strlen (str);
 
-	for (i = 0; i < str_len && !isspace (str[i]); ++i);
+	for (i = 0; i < str_len && !isspace ((int)str[i]); ++i);
 	return i;
 }
